requirement.txt
Django
djangorestframework-simplejwt
